pyproject.toml
[tool.poetry]
name = "fastapi-zero"
version = "0.1.0"
description = ""
authors = ["joaobrc <56743873+joaobrc@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "fastapi_zero"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
pydantic = {extras = ["email"], version = "^2.7.0"}
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.0"
blue = "^0.9.1"
httpx = "^0.24.1"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "0.2.1"


[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.ruff.lint]
preview= true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'uvicorn fastapi_zero.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fastapi_zero -vv'
post_test = 'coverage html'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
